Fixed warning in Linker.cpp.

+++ NULLC/Linker.cpp
@@ -455,7 +455,7 @@ bool Linker::LinkCode(const char *code)
 			cmd.argument = funcRemap[cmd.argument];
 			break;
 		case cmdCloseUpvals:
-			cmd.helper = funcRemap[cmd.helper];
+			cmd.helper = (unsigned short)funcRemap[cmd.helper];
 			cmd.argument += exFunctions[cmd.helper].closeListStart;
 			break;
 		case cmdPushTypeID:
